Add ?new to create a new room automatically
